chore(suite): rename onboarding components, remove ButtonAlt and OptionWrapper
