+++ packages/suite/src/components/onboarding/ConnectDevicePromptManager/ConnectDevicePromptManager.tsx
