+++ packages/suite/src/components/onboarding/DeviceAnimation/DeviceAnimation.tsx
