+++ packages/suite/src/components/onboarding/index.ts
@@ -1,26 +1,23 @@
-import { Option, OptionsWrapper, OptionWrapper, OptionsDivider } from './Option';
+import { OnboardingOption, OptionsWrapper, OptionsDivider } from './OnboardingOption';
 
-import OnboardingButtonAlt from './Buttons/ButtonAlt';
-import OnboardingButtonCta from './Buttons/ButtonCta';
-import OnboardingButtonBack from './Buttons/ButtonBack';
-import OnboardingButtonSkip from './Buttons/ButtonSkip';
+import { OnboardingButtonCta } from './Buttons/OnboardingButtonCta';
+import { OnboardingButtonBack } from './Buttons/OnboardingButtonBack';
+import { OnboardingButtonSkip } from './Buttons/OnboardingButtonSkip';
 
-import { ConnectDevicePromptManager } from './ConnectDevicePromptManager';
+import { ConnectDevicePromptManager } from './ConnectDevicePromptManager/ConnectDevicePromptManager';
 import { OnboardingLayout } from './OnboardingLayout';
 import { Hologram } from './Hologram';
-import { DeviceAnimation, type DeviceAnimationType } from './DeviceAnimation';
+import { DeviceAnimation, type DeviceAnimationType } from './DeviceAnimation/DeviceAnimation';
 import { OnboardingStepBox, type OnboardingStepBoxProps } from './OnboardingStepBox';
-import SkipStepConfirmation from './SkipStepConfirmation';
-import { ProgressBar } from './ProgressBar';
+import { SkipStepConfirmation } from './SkipStepConfirmation';
+import { OnboardingProgressBar } from './OnboardingProgressBar';
 
 export {
-    OnboardingButtonAlt,
     OnboardingButtonCta,
     OnboardingButtonBack,
     OnboardingButtonSkip,
-    Option,
+    OnboardingOption,
     OptionsWrapper,
-    OptionWrapper,
     OptionsDivider,
     ConnectDevicePromptManager,
     OnboardingLayout,
@@ -28,7 +25,7 @@ export {
     DeviceAnimation,
     OnboardingStepBox,
     SkipStepConfirmation,
-    ProgressBar,
+    OnboardingProgressBar,
 };
 
 export type { DeviceAnimationType, OnboardingStepBoxProps };
