+++ packages/suite/src/components/recovery/SelectWordCount.tsx
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
-import { Option, OptionsWrapper, OptionsDivider } from 'src/components/onboarding';
 
+import { OnboardingOption, OptionsWrapper, OptionsDivider } from 'src/components/onboarding';
 import { WordCount } from 'src/types/recovery';
 import { Translation } from 'src/components/suite';
 
-const StyledOption = styled(Option)`
+const StyledOption = styled(OnboardingOption)`
     justify-content: center;
 `;
